split `kubeapiserver` usage in compliance check CLI
