+++ cmd/security-agent/subcommands/check/command.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build !windows && kubeapiserver
+//go:build !windows
 
 // Package check holds check related files
 package check
@@ -14,12 +14,9 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"time"
 
 	"github.com/spf13/cobra"
 	"go.uber.org/fx"
-	"k8s.io/client-go/discovery"
-	"k8s.io/client-go/dynamic"
 
 	ddgostatsd "github.com/DataDog/datadog-go/v5/statsd"
 
@@ -37,10 +34,8 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/compliance/k8sconfig"
 	"github.com/DataDog/datadog-agent/pkg/security/common"
 	"github.com/DataDog/datadog-agent/pkg/security/utils/hostnameutils"
-	"github.com/DataDog/datadog-agent/pkg/util/flavor"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname"
-	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/apiserver"
 )
 
 // CliParams needs to be exported because the compliance subcommand is tightly coupled to this subcommand and tests need to be able to access this type.
@@ -140,20 +135,13 @@ func RunCheck(log log.Component, config config.Component, _ secrets.Component, s
 	var resolver compliance.Resolver
 	if checkArgs.overrideRegoInput != "" {
 		resolver = newFakeResolver(checkArgs.overrideRegoInput)
-	} else if flavor.GetFlavor() == flavor.ClusterAgent {
-		resolver = compliance.NewResolver(context.Background(), compliance.ResolverOptions{
-			Hostname:           hname,
-			DockerProvider:     compliance.DefaultDockerProvider,
-			LinuxAuditProvider: compliance.DefaultLinuxAuditProvider,
-			KubernetesProvider: complianceKubernetesProvider,
-			StatsdClient:       statsdClient,
-		})
 	} else {
 		resolver = compliance.NewResolver(context.Background(), compliance.ResolverOptions{
 			Hostname:           hname,
 			HostRoot:           os.Getenv("HOST_ROOT"),
 			DockerProvider:     compliance.DefaultDockerProvider,
 			LinuxAuditProvider: compliance.DefaultLinuxAuditProvider,
+			KubernetesProvider: complianceKubernetesProvider,
 			StatsdClient:       statsdClient,
 		})
 	}
@@ -259,16 +247,6 @@ func reportComplianceEvents(log log.Component, events []*compliance.CheckEvent, 
 		reporter.ReportEvent(event)
 	}
 	return nil
-}
-
-func complianceKubernetesProvider(_ctx context.Context) (dynamic.Interface, discovery.DiscoveryInterface, error) {
-	ctx, cancel := context.WithTimeout(_ctx, 2*time.Second)
-	defer cancel()
-	apiCl, err := apiserver.WaitForAPIClient(ctx)
-	if err != nil {
-		return nil, nil, err
-	}
-	return apiCl.DynamicCl, apiCl.Cl.Discovery(), nil
 }
 
 type fakeResolver struct {
