+++ cmd/security-agent/subcommands/check/command_unsupported.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build windows || !kubeapiserver
+//go:build windows
 
 // Package check holds check related files
 package check
