+++ cmd/security-agent/subcommands/check/resolver_security_agent.go
@@ -1,0 +1,15 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build !windows && !kubeapiserver
+
+// Package check holds check related files
+package check
+
+import (
+	"github.com/DataDog/datadog-agent/pkg/compliance"
+)
+
+var complianceKubernetesProvider compliance.KubernetesProvider
